Move wrapper utils to a separate file
