+++ src/features/navigator-interface.js
@@ -1,4 +1,5 @@
-import { defineProperty, DDGPromise } from '../utils'
+import { DDGPromise } from '../utils'
+import { defineProperty } from '../wrapper-utils'
 import ContentFeature from '../content-feature'
 
 function injectNavigatorInterface (args) {
