+++ src/utils.js
@@ -239,55 +239,6 @@ export function isFeatureBroken (args, feature) {
         : args.site.isBroken || args.site.allowlisted || !args.site.enabledFeatures.includes(feature)
 }
 
-/**
- * For each property defined on the object, update it with the target value.
- */
-export function overrideProperty (name, prop) {
-    // Don't update if existing value is undefined or null
-    if (!(prop.origValue === undefined)) {
-        /**
-         * When re-defining properties, we bind the overwritten functions to null. This prevents
-         * sites from using toString to see if the function has been overwritten
-         * without this bind call, a site could run something like
-         * `Object.getOwnPropertyDescriptor(Screen.prototype, "availTop").get.toString()` and see
-         * the contents of the function. Appending .bind(null) to the function definition will
-         * have the same toString call return the default [native code]
-         */
-        try {
-            defineProperty(prop.object, name, {
-                // eslint-disable-next-line no-extra-bind
-                get: (() => prop.targetValue).bind(null)
-            })
-        } catch (e) {
-        }
-    }
-    return prop.origValue
-}
-
-export function defineProperty (object, propertyName, descriptor) {
-    if (hasMozProxies) {
-        const usedObj = object.wrappedJSObject
-        const UsedObjectInterface = globalObj.wrappedJSObject.Object
-        const definedDescriptor = new UsedObjectInterface();
-        ['configurable', 'enumerable', 'value', 'writable'].forEach((propertyName) => {
-            if (propertyName in descriptor) {
-                definedDescriptor[propertyName] = cloneInto(
-                    descriptor[propertyName],
-                    definedDescriptor,
-                    { cloneFunctions: true })
-            }
-        });
-        ['get', 'set'].forEach((methodName) => {
-            if (methodName in descriptor) {
-                exportFunction(descriptor[methodName], definedDescriptor, { defineAs: methodName })
-            }
-        })
-        UsedObjectInterface.defineProperty(usedObj, propertyName, definedDescriptor)
-    } else {
-        Object.defineProperty(object, propertyName, descriptor)
-    }
-}
-
 export function camelcase (dashCaseText) {
     return dashCaseText.replace(/-(.)/g, (match, letter) => {
         return letter.toUpperCase()
